+++ src/cli.js
@@ -3,9 +3,21 @@
 const args = require('yargs-parser')(process.argv.slice(2))
 const fs2json = require('./index')
 
-function main () {
+function main() {
 
-  let config = { input: args.input, debug: args.debug, output: args.output }
+  let config = {
+    input: args.input,
+    // debug: args.debug, 
+    output: args.output,
+    // format output json or minified
+    formatted: args.formatted,
+    // custom name for "fileName" property
+    filenamePropertyName: args.filenamePropertyName,
+    // custom name for "content" property
+    contentPropertyName: args.contentPropertyName,
+    //  instead of a map `{[fileName: string]: {fileName: string, content: string, isBinary: boolean}}` output an array of type `{fileName: string, content: string, isBinary: boolean}[]`
+    outputAsArray: args.outputAsArray
+  }
   if (!config.input || !config.output) {
     console.log('Incorrect call, aborting. ')
     process.exit(1)